src/scripts/analyze_image.py
#!/usr/bin/env python
import sys
import json
import random
from PIL import Image

def analyze_image(image_path):
    """
    Mock image analysis function that simulates AI detection of skin conditions.
    For a real implementation, this would use a pretrained Vision Transformer model.
    
    Args:
        image_path: Path to the image file
        
    Returns:
        Dictionary with condition and confidence score
    """
    try:
        # In a real implementation, this would:
        # 1. Load the image
        # 2. Preprocess (resize to 224x224, normalize)
        # 3. Run through the ViT model
        # 4. Return top prediction
        
        # Mock implementation for prototype
        image = Image.open(image_path)
        
        # Mock skin conditions and random confidence scores
        conditions = [
            "eczema",
            "psoriasis", 
            "acne",
            "rosacea",
            "dermatitis",
            "melanoma",
            "rash"
        ]
        
        condition = random.choice(conditions)
        confidence = round(random.uniform(0.7, 0.98), 2)
        
        return {
            "condition": condition,
            "confidence": confidence
        }
        
    except Exception as e:
        print(f"Error analyzing image: {str(e)}", file=sys.stderr)
        return {
            "condition": "unknown",
            "confidence": 0
        }

if __name__ == "__main__":
    if len(sys.argv) != 2:
        print("Usage: python analyze_image.py <image_path>")
        sys.exit(1)
        
    image_path = sys.argv[1]
    result = analyze_image(image_path)
    
    # Print JSON result to stdout for the Node.js process to capture
    print(json.dumps(result)) 